requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
packaging==24.0
PyQt5==5.15.10
python-dotenv==1.0.1
PyQtWebEngine==5.15.6
urllib3==2.1.0
requests==2.31.0
pandas>=1.3.0
matplotlib>=3.4.0
pillow>=8.3.0
sqlalchemy>=1.4.0
bs4==0.0.1
beautifulsoup4==4.12.2
pyqtdarktheme==0.1.7 